* Fixed type error

+++ libcluon/include/cluon/UDPSender.hpp
@@ -96,7 +96,7 @@ class LIBCLUON_API UDPSender {
    private:
     mutable std::mutex m_socketMutex{};
     int32_t m_socket{-1};
-    uint32_t m_portToSentFrom{0};
+    uint16_t m_portToSentFrom{0};
     struct sockaddr_in m_sendToAddress {};
 };
 } // namespace cluon
